+++ src/parser/ast.h
@@ -123,9 +123,9 @@ struct IntLit : public Value {
 };
 
 struct FloatLit : public Value {
-    float val;
+    double val;
 
-    FloatLit(float val_) : val(val_) {}
+    FloatLit(double val_) : val(val_) {}
 };
 
 struct BigintLit : public Value {
@@ -265,7 +265,7 @@ struct SelectStmt : public TreeNode {
 // Semantic value
 struct SemValue {
     int sv_int;
-    float sv_float;
+    double sv_float;
     long long sv_bigint;
     std::string sv_str;
     long long sv_datetime;
